refactor: extract IngestionStagingPipeline from IngestionService
Moves StageVulnerabilitiesAsync, StageVulnerabilityBatchesAsync,
StageAssetInventorySnapshotAsync, StageAssetBatchesAsync,
EnqueueEnrichmentJobsForRunAsync, NormalizeResults, NormalizeAssetSnapshot,
NormalizeAssetSnapshots, Chunk, and AssetBatchStageSummary out of IngestionService
into a focused IngestionStagingPipeline collaborator. Constants AssetBatchSize and
VulnerabilityBatchSize move with them. ProcessAssetsAsync stays in IngestionService
because it calls ProcessStagedAssetsAsync which cannot yet be extracted.
