+++ tests/PatchHound.Tests/Infrastructure/Services/IngestionServicePhase3Tests.cs
@@ -63,6 +63,7 @@ public class IngestionServicePhase3Tests
             remediationDecisionService: null,
             new IngestionLeaseManager(db, NullLogger<IngestionLeaseManager>.Instance),
             new IngestionCheckpointWriter(db),
+            new IngestionStagingPipeline(db, new EnrichmentJobEnqueuer(db, NullLogger<EnrichmentJobEnqueuer>.Instance), new IngestionLeaseManager(db, NullLogger<IngestionLeaseManager>.Instance), new IngestionCheckpointWriter(db)),
             NullLogger<IngestionService>.Instance);
 
         await ingestion.RunExposureDerivationAsync(tenantId, CancellationToken.None);
